~~**Depends on https://github.com/ethereum/go-ethereum/pull/1853 - please review & merge it first.**~~

NOTE: package `crypto/ecies` needs some refactoring love - this patch is not ideal as package `ecies` is a bit messy. Discussed with @fjl and we have several ideas around how to refactor it. Would prefer to make that refactoring in a separate PR so it's easier to review separated from this change.

- [x] Add libsecp256k1 Scalar Multiplication API based on it's ECDH API and use this in the curve interface function `ScalarMult` for this curve.
- [x] Add benchmark of Scalar Multiplication (lib + Go wrapping code)
- [x] Add shared secret test with static values to catch errors when integrating.
- [x] Move Go secp256k1 curve to crypto/secp256k1 so it can be used in crypto/ecies package (needed to run tests with the correct curve - currently tests are run with another curve - P256)
- [x] Fix RLPx test setup to avoid hanging in case of errors returned from crypto libs
- [x] Some boilerplate added to squeeze this curve into current ECIES impl.

Speedup:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkGenSharedKeyS256     6523809       165052        -97.47%
```