@Gustav-Simonsson 2015-10-16
Looks like P2P RLPx tests fails sometimes with this patch - in progress until we figure out why.