@obscuren 2015-10-16
Nice speedup :)