@Gustav-Simonsson 2015-11-26
@obscuren updated, seems to work as expected. the windows build failure is unrelated.