@obscuren 2015-11-27
I think needs some rebasing