@fjl 2015-12-02
@obscuren this is rebased now